refactor: split tools in diferent files and pass to createHubtypeAIAgent finalTools and instructions
