+++ packages/botonic-plugin-ai-agents/src/tools/index.ts
@@ -1,5 +1,5 @@
 import { EXIT_TOOLS } from './exit'
-import { MESSAGE_TOOLS } from './message'
+import { messageResponse } from './message'
 
 export const EXIT_TOOLS_NAMES = EXIT_TOOLS.map(tool => tool.name)
-export const MANDATORY_TOOLS = [...MESSAGE_TOOLS, ...EXIT_TOOLS]
+export const MANDATORY_TOOLS = [messageResponse, ...EXIT_TOOLS]
